+++ crates/storage/store_db/redb.rs
@@ -778,18 +778,6 @@ impl StoreEngine for RedBStore {
         .await
     }
 
-    fn get_chain_config(&self) -> Result<ChainConfig, StoreError> {
-        match self.read_sync(CHAIN_DATA_TABLE, ChainDataIndex::ChainConfig)? {
-            None => Err(StoreError::Custom("Chain config not found".to_string())),
-            Some(bytes) => {
-                let json = String::from_utf8(bytes.value()).map_err(|_| StoreError::DecodeError)?;
-                let chain_config: ChainConfig =
-                    serde_json::from_str(&json).map_err(|_| StoreError::DecodeError)?;
-                Ok(chain_config)
-            }
-        }
-    }
-
     async fn update_earliest_block_number(
         &self,
         block_number: BlockNumber,
@@ -800,6 +788,18 @@ impl StoreEngine for RedBStore {
             block_number.encode_to_vec(),
         )
         .await
+    }
+
+    async fn get_latest_block_number(&self) -> Result<Option<BlockNumber>, StoreError> {
+        match self
+            .read(CHAIN_DATA_TABLE, ChainDataIndex::LatestBlockNumber)
+            .await?
+        {
+            None => Ok(None),
+            Some(ref rlp) => RLPDecode::decode(&rlp.value())
+                .map(Some)
+                .map_err(|_| StoreError::DecodeError),
+        }
     }
 
     async fn get_earliest_block_number(&self) -> Result<Option<BlockNumber>, StoreError> {
@@ -869,18 +869,6 @@ impl StoreEngine for RedBStore {
             block_number.encode_to_vec(),
         )
         .await
-    }
-
-    async fn get_latest_block_number(&self) -> Result<Option<BlockNumber>, StoreError> {
-        match self
-            .read(CHAIN_DATA_TABLE, ChainDataIndex::LatestBlockNumber)
-            .await?
-        {
-            None => Ok(None),
-            Some(ref rlp) => RLPDecode::decode(&rlp.value())
-                .map(Some)
-                .map_err(|_| StoreError::DecodeError),
-        }
     }
 
     async fn update_pending_block_number(