+++ CHANGELOG.md
@@ -4,9 +4,13 @@
 
 ### 2025-07-31
 
+- Cache chain config and latest canonical block header [#3878](https://github.com/lambdaclass/ethrex/pull/3878)
+
 - Batching of transaction hashes sent in a single NewPooledTransactionHashes message [#3912](https://github.com/lambdaclass/ethrex/pull/3912)
+
 - Make `JUMPDEST` blacklist lazily generated on-demand [#3812](https://github.com/lambdaclass/ethrex/pull/3812)
 - Rewrite Blake2 AVX2 implementation (avoid gather instructions and better loop handling).
+
 
 ### 2025-07-30